// src/features/Table/enums/columnsEnum.ts
export enum ColumnsEnum {
    REQUEST_DATE = "Дата заявки",
    INSIDE_NUMBER = "Внутренний №",
    PROFORM_NUMBER = "№ проформы",
    ORDER_NUMBER = "№ заказа",
    CONTAINER_NUMBER = "№ контейнера",
    PRODUCT = "Товар",
    DELIVERY_METHOD = "Способ доставки",
    PROVIDER = "Поставщик",
    IMPORTER = "Импортер",
    CONDITIONS = "Условия поставки",
    DIRECTION = "Направление",
    STORE = "Склад",
    AGENT = "Агент",
    CONTAINER_TYPE = "Тип контейнера",
    PLACE_OF_DISPATCH = "Место отправки",
    LINE = "Линия",
    READY_DATE = "Дата готовности",
    LOADING_DATE = "Дата загрузки",
    ETD = "ETD",
    ETA = "ETA",
    RELEASE = "Релиз",
    BL_SMGS_CMR = "BL/СМГС/CMR",
    TD = "ТД",
    DATE_DO = "Дата ДО",
    PORT = "Порт",
    IS_DS = "Д/С для подачи",
    FRAHT_ACCOUNT = "Фрахтовый счет",
    IS_DOCS = "Документы для подачи",
    DECLARATION_NUMBER = "№ декларации",
    DECLARATION_ISSUE_DATE = "Дата выпуска декларации",
    AVAILIBILITY_IF_OB = "Наличие ОБ",
    ANSWER_OF_OB = "Ответ ОБ",
    EXPEDITOR = "Экспедитор",
    DESTINATION_STATION = "Станция прибытия",
    KM_TO_DIST = "Осталось км до ст. назначения",
    TRAIN_DEPART_DATE = "Дата отправки по ЖД",
    TRAIN_ARRIVE_DATE = "Дата прибытия по ЖД",
    PICKUP = "Автовывоз",
    STORE_ARRIVE_DATE = "Дата прибытия на склад",
    STOCK_PLACE = "Сток Сдачи",
    COMMENT = "Комментарий",
}

export enum ColumnsKeysEnum {
    REQUEST_DATE = "REQUEST_DATE",
    INSIDE_NUMBER = "INSIDE_NUMBER",
    PROFORM_NUMBER = "PROFORM_NUMBER",
    ORDER_NUMBER = "ORDER_NUMBER",
    CONTAINER_NUMBER = "CONTAINER_NUMBER",
    PRODUCT = "SIMPLE_PRODUCT_NAME",
    DELIVERY_METHOD = "DELIVERY_METHOD",
    PROVIDER = "PROVIDERS",
    IMPORTER = "IMPORTERS",
    CONDITIONS = "CONDITIONS",
    DIRECTION = "DIRECTION",
    STORE = "STORE_NAME",
    AGENT = "AGENT",
    CONTAINER_TYPE = "CONTAINER_TYPE",
    PLACE_OF_DISPATCH = "PLACE_OF_DISPATCH",
    LINE = "LINE",
    READY_DATE = "READY_DATE",
    LOADING_DATE = "LOADING_DATE",
    ETD = "ETD",
    ETA = "ETA",
    RELEASE = "RELEASE",
    BL_SMGS_CMR = "BL_SMGS_CMR",
    TD = "TD",
    DATE_DO = "DATE_DO",
    PORT = "PORT",
    IS_DS = "IS_DS",
    FRAHT_ACCOUNT = "FRAHT_ACCOUNT",
    IS_DOCS = "IS_DOCS",
    DECLARATION_NUMBER = "DECLARATION_NUMBER",
    DECLARATION_ISSUE_DATE = "DECLARATION_ISSUE_DATE",
    AVAILIBILITY_IF_OB = "AVAILIBILITY_IF_OB",
    ANSWER_OF_OB = "ANSWER_OF_OB",
    EXPEDITOR = "EXPEDITOR",
    DESTINATION_STATION = "DESTINATION_STATION",
    KM_TO_DIST = "KM_TO_DIST",
    TRAIN_DEPART_DATE = "TRAIN_DEPART_DATE",
    TRAIN_ARRIVE_DATE = "TRAIN_ARRIVE_DATE",
    PICKUP = "PICKUP",
    STORE_ARRIVE_DATE = "STORE_ARRIVE_DATE",
    STOCK_PLACE = "STOCK_PLACE",
    COMMENT = "COMMENT",
}